// src/core/api/generated/spring/schemas/MeetingReminderDTOStatus.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Crm Backend API
 * Crm Backend API documentation
 * OpenAPI spec version: 0.0.1
 */

export type MeetingReminderDTOStatus = typeof MeetingReminderDTOStatus[keyof typeof MeetingReminderDTOStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MeetingReminderDTOStatus = {
  DRAFT: 'DRAFT',
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
  ARCHIVED: 'ARCHIVED',
} as const;